pyproject.toml
[tool.poetry]
name = "real-time-chat"
version = "0.1.0"
description = ""
authors = ["sendervision <sendervision@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-socketio = "^5.11.3"
flask = "^3.0.3"
flask-socketio = "^5.3.6"
flask-sqlalchemy = "^3.1.1"
python-dotenv = "^1.0.1"
pydantic = "^2.7.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
